requirements.txt

streamlit
opencv-python
numpy
Pillow
fpdf
pytesseract
